requirements.txt
asyncio
discord.py
python-dotenv
googletrans==3.1.0a0
pynacl
